add edit product dialog
