+++ FinancialAssistantApp/app/src/main/java/com/coderaptor/financial/assistant/app/ReceiptActivity.kt
@@ -4,13 +4,8 @@ import android.content.Intent
 import android.os.Bundle
 import android.text.Editable
 import android.util.Log
-import android.view.View
-import android.widget.AdapterView
 import android.widget.ArrayAdapter
 import androidx.appcompat.app.AppCompatActivity
-import com.afollestad.materialdialogs.MaterialDialog
-import com.afollestad.materialdialogs.customview.customView
-import com.afollestad.materialdialogs.customview.getCustomView
 import com.afollestad.recyclical.datasource.DataSource
 import com.afollestad.recyclical.datasource.emptyDataSourceTyped
 import com.afollestad.recyclical.setup
@@ -21,15 +16,17 @@ import com.coderaptor.financial.assistant.app.adapters.ProductViewHolder
 import com.coderaptor.financial.assistant.app.core.Product
 import com.coderaptor.financial.assistant.app.core.Receipt
 import com.coderaptor.financial.assistant.app.data.DatabaseHandler
-import com.coderaptor.financial.assistant.app.util.*
+import com.coderaptor.financial.assistant.app.gui.dialogs.openProductDialog
+import com.coderaptor.financial.assistant.app.util.fieldsEmpty
+import com.coderaptor.financial.assistant.app.util.formatDate
+import com.coderaptor.financial.assistant.app.util.openCalendar
 import com.coderaptor.financial.assistant.app.util.spinner.StringWithId
 import com.coderaptor.financial.assistant.app.util.spinner.getCategoryStringWithIdList
-import com.google.android.material.snackbar.Snackbar
+import com.coderaptor.financial.assistant.app.util.toast
 import kotlinx.android.synthetic.main.activity_receipt.*
 import kotlinx.android.synthetic.main.content_receipt.*
-import kotlinx.android.synthetic.main.content_receipt.dateField
 import kotlinx.android.synthetic.main.dialog_add_income.view.*
-import kotlinx.android.synthetic.main.dialog_add_product.*
+import java.util.*
 
 class ReceiptActivity: AppCompatActivity() {
     val dbHandler = DatabaseHandler(this)
@@ -135,77 +132,7 @@ class ReceiptActivity: AppCompatActivity() {
         adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
 
         fab.setOnClickListener {
-            MaterialDialog(this).show {
-                setTheme(R.style.AppTheme)
-                title(R.string.product_string)
-                customView(R.layout.dialog_add_product, scrollable = true)
-
-                categoryField.adapter = adapter
-
-                var categoryId = (-1).toLong()
-                val categoryIdWithWarrantyAndHarmful = dbHandler.categoriesWithWarrantyAndHarmful()
-                categoryField.onItemSelectedListener = object : AdapterView.OnItemSelectedListener {
-                    override fun onItemSelected(parent: AdapterView<*>, view: View, position: Int, id: Long) {
-                        val swt = parent.selectedItem as StringWithId
-                        categoryId = swt.id
-                        if (categoryIdWithWarrantyAndHarmful.contains(categoryId) && !dbHandler.isHarmfulCategory(categoryId)) {
-                            dateField.visibility = View.VISIBLE
-                            label.visibility = View.VISIBLE
-                            warranty.isChecked = true
-                        }else if (!categoryIdWithWarrantyAndHarmful.contains(categoryId)) {
-                            dateField.visibility = View.INVISIBLE
-                            label.visibility = View.INVISIBLE
-                            warranty.isChecked = false
-                        }else if (categoryIdWithWarrantyAndHarmful.contains(categoryId) && dbHandler.isHarmfulCategory(categoryId)) {
-                            if (SharedPreference.shoppingMonitor) {
-                                Snackbar.make(view, "Káros Termék", Snackbar.LENGTH_SHORT)
-                                    .setAction("Action", null).show()
-                            }
-                        }
-                    }
-                    override fun onNothingSelected(parent: AdapterView<*>) {}
-                }
-
-                val datefield = getCustomView().dateField
-                datefield.isClickable = true
-                datefield.text = Editable.Factory.getInstance().newEditable(java.util.Calendar.getInstance().formatDate())
-                datefield.setOnClickListener {
-                    openCalendar(datefield)
-                }
-
-                warranty.setOnCheckedChangeListener { _, isChecked ->
-                    if(isChecked) {
-                        dateField.visibility = View.VISIBLE
-                        label.visibility = View.VISIBLE
-                    }else {
-                        dateField.visibility = View.INVISIBLE
-                        label.visibility = View.INVISIBLE
-                    }
-                }
-
-                positiveButton(R.string.save) {
-                    val result = fieldsEmpty(productName.text, quantityField.text, priceField.text)
-                    if (result) {
-                        val name = productName.text.toString()
-                        val quantity = if (quantityField.text.isNotEmpty()) quantityField.text.toString().toInt() else 1
-                        val unit = unitField.selectedItem.toString()
-                        val price = priceField.text.toString().toInt()
-                        val warranty = warranty.isChecked
-                        var product = Product(name, unit, quantity, price, categoryId = categoryId)
-                        if (warranty) {
-                            if (dateField.text.isNotEmpty()) {
-                                val endDate = dateField.text.toString()
-                                product = Product(name, unit, quantity, price, endDate, categoryId)
-                            }
-                        }
-                        dataSource.add(product)
-                        toast("Sikeres hozzáadás")
-                    } else {
-                        toast("Hiányzó adat!")
-                    }
-                }
-                negativeButton(R.string.cancel)
-            }
+            openProductDialog(dataSource, dbHandler, adapter = adapter)
         }
 
         recyclerView.setup {
@@ -215,7 +142,6 @@ class ReceiptActivity: AppCompatActivity() {
                 text(R.string.delete)
                 color(R.color.delete)
                 callback { index, item ->
-                    //toast("delete $index: $item")
                     if (item is Product)
                         dbHandler.deleteByPosition(item.id, DatabaseHandler.TABLE_NAME_PRODUCT)
                     true
@@ -227,9 +153,8 @@ class ReceiptActivity: AppCompatActivity() {
                 text(R.string.edit)
                 color(R.color.edit)
                 callback { index, item ->
-                    //toast("edit $index: $item")
                     if (item is Product) {
-                        //edit layout
+                        openProductDialog(dataSource, dbHandler, item, adapter)
                     }
                     false
                 }
